#18: integration of buildTemplate into createDeltaPkg
