+++ lib/Builder.js
@@ -3,6 +3,7 @@
 const TYPE = require('../types/mcdev.d');
 const Util = require('./util/util');
 const File = require('./util/file');
+const Cli = require('./util/cli');
 
 const MetadataTypeInfo = require('./MetadataTypeInfo');
 // @ts-ignore
@@ -100,6 +101,35 @@ class Builder {
         return this.metadata;
     }
     /**
+     * Build a specific metadata file based on a template.
+     *
+     * @param {string} businessUnit references credentials from properties.json
+     * @param {string} selectedType supported metadata type
+     * @param {string} key customerkey of the metadata
+     * @param {string} market market localizations
+     * @returns {Promise.<TYPE.MultiMetadataTypeList>} -
+     */
+    static async buildTemplate(businessUnit, selectedType, key, market) {
+        Util.logger.info('mcdev:: Build Definition from Template');
+        const properties = File.loadConfigFile();
+        if (!Util._isValidType(selectedType)) {
+            return;
+        }
+        if (selectedType.includes('-')) {
+            Util.logger.error(
+                `:: '${selectedType}' is not a valid metadata type. Please don't include subtypes.`
+            );
+            return;
+        }
+        const buObject = await Cli.getCredentialObject(properties, businessUnit);
+        if (buObject !== null) {
+            const builder = new Builder(properties, buObject, null);
+            if (this._checkMarket(market)) {
+                return builder.buildTemplate(selectedType, key, properties.markets[market]);
+            }
+        }
+    }
+    /**
      * Builds a specific metadata file by name
      *
      * @param {string} metadataType metadata type to create a template of
@@ -149,6 +179,91 @@ class Builder {
         return this.metadata;
     }
     /**
+     * Build a specific metadata file based on a template.
+     *
+     * @param {string} businessUnit references credentials from properties.json
+     * @param {string} selectedType supported metadata type
+     * @param {string} name name of the metadata
+     * @param {string} market market localizations
+     * @returns {Promise.<void>} -
+     */
+    static async buildDefinition(businessUnit, selectedType, name, market) {
+        const properties = File.loadConfigFile();
+        if (!Util._isValidType(selectedType)) {
+            return;
+        }
+        if (selectedType.includes('-')) {
+            Util.logger.error(
+                `:: '${selectedType}' is not a valid metadata type. Please don't include subtypes.`
+            );
+            return;
+        }
+        const buObject = await Cli.getCredentialObject(properties, businessUnit);
+        if (buObject !== null) {
+            const builder = new Builder(properties, buObject, null);
+            if (market === '*') {
+                for (const oneMarket in properties.markets) {
+                    builder.buildDefinition(selectedType, name, properties.markets[oneMarket]);
+                }
+            } else {
+                if (this._checkMarket(market)) {
+                    builder.buildDefinition(selectedType, name, properties.markets[market]);
+                }
+            }
+        }
+    }
+    /**
+     * Build a specific metadata file based on a template using a list of bu-market combos
+     *
+     * @param {string} listName name of list of BU-market combos
+     * @param {string} type supported metadata type
+     * @param {string} name name of the metadata
+     * @returns {Promise.<void>} -
+     */
+    static async buildDefinitionBulk(listName, type, name) {
+        const properties = File.loadConfigFile();
+        if (!properties.marketList) {
+            Util.logger.error('Please define properties.marketList object in your config');
+            return;
+        }
+        if (!properties.marketList[listName]) {
+            Util.logger.error(`Please define properties.marketList.${listName} in your config`);
+            return;
+        }
+        if (type && !MetadataTypeInfo[type]) {
+            Util.logger.error(`:: '${type}' is not a valid metadata type`);
+            return;
+        }
+        let i = 0;
+        for (const businessUnit in properties.marketList[listName]) {
+            if (businessUnit === 'description') {
+                // skip, it's just a metadata on this list and not a BU
+                continue;
+            }
+            i++;
+            const market = properties.marketList[listName][businessUnit];
+            let marketList = [];
+            if ('string' === typeof market) {
+                marketList.push(market);
+            } else {
+                marketList = market;
+            }
+            marketList.forEach((market) => {
+                if (market && properties.markets[market]) {
+                    Util.logger.info(`Executing for '${businessUnit}': '${market}'`);
+                    this.buildDefinition(businessUnit, type, name, market);
+                } else {
+                    Util.logger.error(
+                        `Could not find '${market}' in properties.markets. Please check your properties.marketList.${listName} confguration.`
+                    );
+                }
+            });
+        }
+        if (!i) {
+            Util.logger.error('Please define properties.marketList in your config');
+        }
+    }
+    /**
      * ensure provided MarketList exists and it's content including markets and BUs checks out
      *
      * @param {string} mlName name of marketList
@@ -194,6 +309,28 @@ class Builder {
             if (!buCounter) {
                 throw new Error(`No BUs defined in marketList ${mlName}`);
             }
+        }
+    }
+    /**
+     * check if a market name exists in current mcdev config
+     *
+     * @param {string} market market localizations
+     * @returns {boolean} found market or not
+     */
+    static _checkMarket(market) {
+        const properties = File.loadConfigFile();
+        if (properties.markets[market]) {
+            return true;
+        } else {
+            Util.logger.error(`Could not find the market '${market}' in your configuration file.`);
+            const marketArr = [];
+            for (const oneMarket in properties.markets) {
+                marketArr.push(oneMarket);
+            }
+            if (marketArr.length) {
+                Util.logger.info('Available markets are: ' + marketArr.join(', '));
+            }
+            return false;
         }
     }
 }
