+++ lib/index.js
@@ -15,7 +15,6 @@ const Deployer = require('./Deployer');
 const MetadataTypeInfo = require('./MetadataTypeInfo');
 const MetadataTypeDefinitions = require('./MetadataTypeDefinitions');
 const Retriever = require('./Retriever');
-const inquirer = require('inquirer');
 const cache = require('./util/cache');
 let properties;
 
@@ -44,7 +43,7 @@ class Mcdev {
      * @param {string} [argv.range] git commit range
     into deploy directory
      * @param {string} [argv.filter] filter file paths that start with any
-     * @param {boolean} [argv.skipInteraction] allows to skip interactive wizard
+     * @param {object} [argv.skipInteraction] allows to skip interactive wizard
      * @returns {void}
      */
     static async createDeltaPkg(argv) {
@@ -54,188 +53,10 @@ class Mcdev {
         }
         // get source market and source BU from config
         if (argv.filter) {
-            return DevOps.createDeltaPkg(properties, argv.range, true, argv.filter);
+            return DevOps.getDeltaList(properties, argv.range, true, argv.filter);
         } else {
             // If no custom filter was provided, use deployment marketLists & templating
-
-            // check if sourceTargetMapping is valid
-            if (
-                !properties.options.deployment.sourceTargetMapping ||
-                !Object.keys(properties.options.deployment.sourceTargetMapping).length
-            ) {
-                Util.logger.error('Bad configuration of options.deployment.sourceTargetMapping');
-                return;
-            }
-            const sourceMarketListArr = Object.keys(
-                properties.options.deployment.sourceTargetMapping
-            );
-
-            for (const sourceML of sourceMarketListArr) {
-                // check if sourceTargetMapping has valid values
-                // #1 check source marketlist
-                try {
-                    Builder.verifyMarketList(sourceML, properties);
-                    // remove potentially existing "description"-entry
-                    delete properties.marketList[sourceML].description;
-
-                    const sourceMarketBuArr = Object.keys(properties.marketList[sourceML]);
-                    if (sourceMarketBuArr.length !== 1) {
-                        throw new Error('Only 1 BU is allowed per source marketList');
-                    }
-                    if ('string' !== typeof properties.marketList[sourceML][sourceMarketBuArr[0]]) {
-                        throw new Error('Only 1 market per BU is allowed per source marketList');
-                    }
-                } catch (ex) {
-                    Util.logger.error('Deployment Source: ' + ex.message);
-                    return;
-                }
-                // #2 check corresponding target marketList
-                let targetML;
-                try {
-                    targetML = properties.options.deployment.sourceTargetMapping[sourceML];
-                    if ('string' !== typeof targetML) {
-                        throw new Error(
-                            'Please define one target marketList per source in deployment.sourceTargetMapping (No arrays allowed)'
-                        );
-                    }
-                    Builder.verifyMarketList(targetML, properties);
-                    // remove potentially existing "description"-entry
-                    delete properties.marketList[targetML].description;
-                } catch (ex) {
-                    Util.logger.error('Deployment Target: ' + ex.message);
-                }
-            }
-            // all good let's loop a second time for actual execution
-            for (const sourceMlName of sourceMarketListArr) {
-                const targetMlName =
-                    properties.options.deployment.sourceTargetMapping[sourceMlName];
-                const sourceBU = Object.keys(properties.marketList[sourceMlName])[0];
-                const sourceMarket = Object.values(properties.marketList[sourceMlName])[0];
-
-                const delta = await DevOps.createDeltaPkg(properties, argv.range, false, sourceBU);
-                // If only chaing templating and buildDefinition if required
-                if (!delta || delta.length === 0) {
-                    // info/error messages was printed by DevOps.createDeltaPkg() already
-                    return;
-                }
-                Util.logger.info('=============');
-
-                // Put files into maps. One map with BU -> type -> file (for retrieveAsTemplate)
-                // Other map only with type -> file (for buildDefinitionBulk)
-                const buTypeDelta = {};
-                const typeDelta = {};
-                let deltaCounter = 0;
-                delta
-                    // Only template/build files that were added/updated/moved. no deletions
-                    // ! doesn't work for folder, because their name parsing doesnt work at the moment
-                    .filter((file) => file.gitAction !== 'delete' && file.name)
-                    .forEach((file) => {
-                        const buPath = `${file._credential}/${file._businessUnit}`;
-                        if (!buTypeDelta[buPath]) {
-                            // init object
-                            buTypeDelta[buPath] = {};
-                        }
-                        if (!buTypeDelta[buPath][file.type]) {
-                            // init array
-                            buTypeDelta[buPath][file.type] = [];
-                        }
-                        buTypeDelta[buPath][file.type].push(file.name);
-                    });
-
-                // Run retrieve as template for each business unit for each type
-                Util.logger.info('Retrieve template from Git delta');
-                // ! needs to be for (.. in ..) loop so that it gets executed in series
-                for (const bu in buTypeDelta) {
-                    for (const type in buTypeDelta[bu]) {
-                        Util.logger.info(
-                            `⚡ mcdev rt ${bu} ${type} "${buTypeDelta[bu][type].join(
-                                ','
-                            )}" ${sourceMarket}`
-                        );
-                        const result = await Mcdev.retrieveAsTemplate(
-                            bu,
-                            type,
-                            buTypeDelta[bu][type].join(','),
-                            sourceMarket
-                        );
-                        // ensure we have the right key for bd/bdb that matches the name used for rt
-                        const keys = result[type]
-                            .map((item) => item[MetadataTypeInfo[type].definition.keyField])
-                            .filter((item) => !!item);
-                        if (keys.length) {
-                            if (!typeDelta[type]) {
-                                // init array
-                                typeDelta[type] = [];
-                            }
-                            typeDelta[type].push(...keys);
-                            deltaCounter += keys.length;
-                        }
-                    }
-                }
-                if (deltaCounter) {
-                    Util.logger.info(`- ✔️  Templates created: ${deltaCounter}`);
-                } else {
-                    Util.logger.error(
-                        '- ❌ No Templates or Deploy Definitions created. Check if you expected no changes.'
-                    );
-                    return;
-                }
-
-                // Run build definitions bulk for each type
-                Util.logger.info('=============');
-                Util.logger.info('Build deploy definitions from delta templates');
-                if (
-                    properties.directories.templateBuilds == properties.directories.deploy ||
-                    (Array.isArray(properties.directories.templateBuilds) &&
-                        properties.directories.templateBuilds.includes(
-                            properties.directories.deploy
-                        ))
-                ) {
-                    let responses;
-                    if (!argv.skipInteraction) {
-                        // deploy folder is in targets for definition creation
-                        // recommend to purge their content first
-                        responses = await inquirer.prompt([
-                            {
-                                type: 'confirm',
-                                name: 'isPurgeDeployFolder',
-                                message:
-                                    'Do you want to empty the deploy folder (ensures no files from previous deployments remain)?',
-                                default: true,
-                            },
-                        ]);
-                    }
-                    if (argv.skipInteraction || responses.isPurgeDeployFolders) {
-                        // Clear output folder structure for selected sub-type
-                        File.removeSync(File.normalizePath([properties.directories.deploy]));
-                    }
-                }
-                const bdPromises = [];
-                for (const type in typeDelta) {
-                    Util.logger.info(
-                        `⚡ mcdev bdb ${targetMlName} ${type} "${typeDelta[type].join(',')}"`
-                    );
-                    // omitting "await" to speed up creation
-                    bdPromises.push(
-                        this.buildDefinitionBulk(targetMlName, type, typeDelta[type].join(','))
-                    );
-                }
-                await Promise.all(bdPromises);
-                Util.logger.info(`- ✔️  Deploy defintions created`);
-                if (
-                    properties.directories.templateBuilds == properties.directories.deploy ||
-                    (Array.isArray(properties.directories.templateBuilds) &&
-                        properties.directories.templateBuilds.includes(
-                            properties.directories.deploy
-                        ))
-                ) {
-                    Util.logger.info(`You can now run deploy on the prepared BUs`);
-                } else {
-                    Util.logger.info(
-                        `Your templated defintions are now ready to be copied into the deploy folder. Hint: You can have this auto-copied if you adjust directories.templateBuilds in your config.`
-                    );
-                }
-            }
+            return DevOps.buildDeltaDefinitions(properties, argv.range, argv.skipInteraction);
         }
     }
 
@@ -776,7 +597,7 @@ class Mcdev {
                 return;
             }
             const retriever = new Retriever(properties, buObject, client);
-            if (this._checkMarket(market)) {
+            if (Builder._checkMarket(market)) {
                 return retriever.retrieve(retrieveTypesArr, name, properties.markets[market]);
             }
         }
@@ -793,23 +614,7 @@ class Mcdev {
      */
     static async buildTemplate(businessUnit, selectedType, key, market) {
         Util.logger.info('mcdev:: Build Template from retrieve folder');
-        properties = properties || File.loadConfigFile();
-        if (!Util._isValidType(selectedType)) {
-            return;
-        }
-        if (selectedType.includes('-')) {
-            Util.logger.error(
-                `:: '${selectedType}' is not a valid metadata type. Please don't include subtypes.`
-            );
-            return;
-        }
-        const buObject = await Cli.getCredentialObject(properties, businessUnit);
-        if (buObject !== null) {
-            const builder = new Builder(properties, buObject, null);
-            if (this._checkMarket(market)) {
-                return builder.buildTemplate(selectedType, key, properties.markets[market]);
-            }
-        }
+        return Builder.buildTemplate(businessUnit, selectedType, key, market);
     }
     /**
      * Build a specific metadata file based on a template.
@@ -822,51 +627,7 @@ class Mcdev {
      */
     static async buildDefinition(businessUnit, selectedType, name, market) {
         Util.logger.info('mcdev:: Build Definition from Template');
-        properties = properties || File.loadConfigFile();
-        if (!Util._isValidType(selectedType)) {
-            return;
-        }
-        if (selectedType.includes('-')) {
-            Util.logger.error(
-                `:: '${selectedType}' is not a valid metadata type. Please don't include subtypes.`
-            );
-            return;
-        }
-        const buObject = await Cli.getCredentialObject(properties, businessUnit);
-        if (buObject !== null) {
-            const builder = new Builder(properties, buObject, null);
-            if (market === '*') {
-                for (const oneMarket in properties.markets) {
-                    builder.buildDefinition(selectedType, name, properties.markets[oneMarket]);
-                }
-            } else {
-                if (this._checkMarket(market)) {
-                    builder.buildDefinition(selectedType, name, properties.markets[market]);
-                }
-            }
-        }
-    }
-    /**
-     * check if a market name exists in current mcdev config
-     *
-     * @param {string} market market localizations
-     * @returns {boolean} found market or not
-     */
-    static _checkMarket(market) {
-        properties = properties || File.loadConfigFile();
-        if (properties.markets[market]) {
-            return true;
-        } else {
-            Util.logger.error(`Could not find the market '${market}' in your configuration file.`);
-            const marketArr = [];
-            for (const oneMarket in properties.markets) {
-                marketArr.push(oneMarket);
-            }
-            if (marketArr.length) {
-                Util.logger.info('Available markets are: ' + marketArr.join(', '));
-            }
-            return false;
-        }
+        Builder.buildDefinition(businessUnit, selectedType, name, market);
     }
 
     /**
@@ -879,47 +640,7 @@ class Mcdev {
      */
     static async buildDefinitionBulk(listName, type, name) {
         Util.logger.info('mcdev:: Build Definition from Template Bulk');
-        properties = properties || File.loadConfigFile();
-        if (!properties.marketList) {
-            Util.logger.error('Please define properties.marketList object in your config');
-            return;
-        }
-        if (!properties.marketList[listName]) {
-            Util.logger.error(`Please define properties.marketList.${listName} in your config`);
-            return;
-        }
-        if (type && !MetadataTypeInfo[type]) {
-            Util.logger.error(`:: '${type}' is not a valid metadata type`);
-            return;
-        }
-        let i = 0;
-        for (const businessUnit in properties.marketList[listName]) {
-            if (businessUnit === 'description') {
-                // skip, it's just a metadata on this list and not a BU
-                continue;
-            }
-            i++;
-            const market = properties.marketList[listName][businessUnit];
-            let marketList = [];
-            if ('string' === typeof market) {
-                marketList.push(market);
-            } else {
-                marketList = market;
-            }
-            marketList.forEach((market) => {
-                if (market && properties.markets[market]) {
-                    Util.logger.info(`Executing for '${businessUnit}': '${market}'`);
-                    this.buildDefinition(businessUnit, type, name, market);
-                } else {
-                    Util.logger.error(
-                        `Could not find '${market}' in properties.markets. Please check your properties.marketList.${listName} confguration.`
-                    );
-                }
-            });
-        }
-        if (!i) {
-            Util.logger.error('Please define properties.marketList in your config');
-        }
+        Builder.buildDefinitionBulk(listName, type, name);
     }
 }
