+++ lib/util/devops.js
@@ -1,8 +1,10 @@
+const TYPE = require('../../types/mcdev.d');
 const File = require('./file');
 const path = require('path');
 const inquirer = require('inquirer');
 const Util = require('./util');
 const git = require('simple-git')();
+const Builder = require('../Builder');
 const MetadataType = require('../MetadataTypeInfo');
 const jsonToTable = require('json-to-table');
 /**
@@ -18,9 +20,9 @@ const DevOps = {
      * @param {string} [range] git commit range
      * @param {boolean} [saveToDeployDir] if true, copy metadata changes into deploy directory
      * @param {string} [filterPaths] filter file paths that start with any specified path (comma separated)
-     * @returns {Promise.<object[]>} -
+     * @returns {Promise.<TYPE.DeltaPkgItem[]>} -
      */
-    async createDeltaPkg(properties, range, saveToDeployDir, filterPaths) {
+    async getDeltaList(properties, range, saveToDeployDir, filterPaths) {
         const rangeUserInput = range;
         Util.logger.info('Create Delta Package ::');
         if (filterPaths) {
@@ -74,9 +76,12 @@ const DevOps = {
             'add/update': 0,
             move: 0,
         };
+        /**
+         * @type {TYPE.DeltaPkgItem[]}
+         */
         const delta = (await git.diffSummary([range])).files
             // populate additional info for all changed files
-            .map((file) => {
+            .map((/** @type {TYPE.DeltaPkgItem} */ file) => {
                 // If file was moved it's path needs to be parsed
                 file.moved = file.file.includes('=>');
                 if (file.moved) {
@@ -109,7 +114,7 @@ const DevOps = {
             // ! Filter happens after initial parse, because if file was moved its new path has to be calculated
             .filter((file) => filterPaths.some((path) => file.file.startsWith(path)))
             // ensure badly named files on unsupported metadata types are not in our subset
-            .filter((file) => {
+            .filter((/** @type {TYPE.DeltaPkgItem} */ file) => {
                 if (!MetadataType[file.type]) {
                     Util.logger.debug(
                         `Unknown metadata-type found for (${file.file}): ` + file.type
@@ -119,7 +124,7 @@ const DevOps = {
                     return true;
                 }
             })
-            .map((file) => {
+            .map((/** @type {TYPE.DeltaPkgItem} */ file) => {
                 // Gets external key based on file name und the assumption that filename = externalKey
                 if (file.type === 'folder') {
                     file.externalKey = null;
@@ -186,7 +191,7 @@ const DevOps = {
         }
         // Write into delta.json to serve as documentation
         File.writeJSONToFile(properties.directories.deltaPackage, 'delta_package', delta);
-        this.document(properties.directories.deltaPackage, delta);
+        DevOps.document(properties.directories.deltaPackage, delta);
         Util.logger.info(
             `- ✔️  Identified changes: Add/Update=${gitActionsCounter['add/update']}, Move=${gitActionsCounter['move']}, Delete=${gitActionsCounter['delete']}`
         );
@@ -201,6 +206,7 @@ const DevOps = {
 
         // Copy filtered list of files into deploy directory
         if (saveToDeployDir) {
+            /** @type {TYPE.DeltaPkgItem} */
             const copied = delta.map((file) =>
                 File.copyFile(
                     file.file,
@@ -237,6 +243,187 @@ const DevOps = {
         }
         return delta;
     },
+    /**
+     * wrapper around DevOps.getDeltaList, Builder.buildTemplate and M
+     *
+     * @param {object} properties project config file
+     * @param {string} range git commit range
+     * @param {TYPE.SkipInteraction} [skipInteraction] allows to skip interactive wizard
+     */
+    async buildDeltaDefinitions(properties, range, skipInteraction) {
+        // check if sourceTargetMapping is valid
+        if (
+            !properties.options.deployment.sourceTargetMapping ||
+            !Object.keys(properties.options.deployment.sourceTargetMapping).length
+        ) {
+            Util.logger.error('Bad configuration of options.deployment.sourceTargetMapping');
+            return;
+        }
+        const sourceMarketListArr = Object.keys(properties.options.deployment.sourceTargetMapping);
+
+        for (const sourceML of sourceMarketListArr) {
+            // check if sourceTargetMapping has valid values
+            // #1 check source marketlist
+            try {
+                Builder.verifyMarketList(sourceML, properties);
+                // remove potentially existing "description"-entry
+                delete properties.marketList[sourceML].description;
+
+                const sourceMarketBuArr = Object.keys(properties.marketList[sourceML]);
+                if (sourceMarketBuArr.length !== 1) {
+                    throw new Error('Only 1 BU is allowed per source marketList');
+                }
+                if ('string' !== typeof properties.marketList[sourceML][sourceMarketBuArr[0]]) {
+                    throw new Error('Only 1 market per BU is allowed per source marketList');
+                }
+            } catch (ex) {
+                Util.logger.error('Deployment Source: ' + ex.message);
+                return;
+            }
+            // #2 check corresponding target marketList
+            let targetML;
+            try {
+                targetML = properties.options.deployment.sourceTargetMapping[sourceML];
+                if ('string' !== typeof targetML) {
+                    throw new Error(
+                        'Please define one target marketList per source in deployment.sourceTargetMapping (No arrays allowed)'
+                    );
+                }
+                Builder.verifyMarketList(targetML, properties);
+                // remove potentially existing "description"-entry
+                delete properties.marketList[targetML].description;
+            } catch (ex) {
+                Util.logger.error('Deployment Target: ' + ex.message);
+            }
+        }
+        // all good let's loop a second time for actual execution
+        for (const sourceMlName of sourceMarketListArr) {
+            const targetMlName = properties.options.deployment.sourceTargetMapping[sourceMlName];
+            const sourceBU = Object.keys(properties.marketList[sourceMlName])[0];
+            const sourceMarket = Object.values(properties.marketList[sourceMlName])[0];
+
+            const delta = await DevOps.getDeltaList(properties, range, false, sourceBU);
+            // If only chaing templating and buildDefinition if required
+            if (!delta || delta.length === 0) {
+                // info/error messages was printed by DevOps.createDeltaPkg() already
+                return;
+            }
+            Util.logger.info('=============');
+
+            // Put files into maps. One map with BU -> type -> file (for retrieveAsTemplate)
+            // Other map only with type -> file (for buildDefinitionBulk)
+            const buTypeDelta = {};
+            const typeDelta = {};
+            let deltaCounter = 0;
+            delta
+                // Only template/build files that were added/updated/moved. no deletions
+                // ! doesn't work for folder, because their name parsing doesnt work at the moment
+                .filter((file) => file.gitAction !== 'delete' && file.name)
+                .forEach((file) => {
+                    const buPath = `${file._credential}/${file._businessUnit}`;
+                    if (!buTypeDelta[buPath]) {
+                        // init object
+                        buTypeDelta[buPath] = {};
+                    }
+                    if (!buTypeDelta[buPath][file.type]) {
+                        // init array
+                        buTypeDelta[buPath][file.type] = [];
+                    }
+                    buTypeDelta[buPath][file.type].push(file.key);
+                });
+
+            // Run buildTemplate for each business unit for each type
+            Util.logger.info('Retrieve template from Git delta');
+            // ! needs to be for (.. in ..) loop so that it gets executed in series
+            for (const bu in buTypeDelta) {
+                for (const type in buTypeDelta[bu]) {
+                    Util.logger.info(
+                        `⚡ mcdev rt ${bu} ${type} "${buTypeDelta[bu][type].join(
+                            ','
+                        )}" ${sourceMarket}`
+                    );
+                    const result = await Builder.buildTemplate(
+                        bu,
+                        type,
+                        buTypeDelta[bu][type].join(','),
+                        sourceMarket
+                    );
+                    // ensure we have the right key for bd/bdb that matches the name used for rt
+                    const keys = result[type]
+                        .map((item) => item[MetadataType[type].definition.keyField])
+                        .filter((item) => !!item);
+                    if (keys.length) {
+                        if (!typeDelta[type]) {
+                            // init array
+                            typeDelta[type] = [];
+                        }
+                        typeDelta[type].push(...keys);
+                        deltaCounter += keys.length;
+                    }
+                }
+            }
+            if (deltaCounter) {
+                Util.logger.info(`- ✔️  Templates created: ${deltaCounter}`);
+            } else {
+                Util.logger.error(
+                    '- ❌ No Templates or Deploy Definitions created. Check if you expected no changes.'
+                );
+                return;
+            }
+
+            // Run build definitions bulk for each type
+            Util.logger.info('=============');
+            Util.logger.info('Build deploy definitions from delta templates');
+            if (
+                properties.directories.templateBuilds == properties.directories.deploy ||
+                (Array.isArray(properties.directories.templateBuilds) &&
+                    properties.directories.templateBuilds.includes(properties.directories.deploy))
+            ) {
+                let responses;
+                if (!skipInteraction) {
+                    // deploy folder is in targets for definition creation
+                    // recommend to purge their content first
+                    responses = await inquirer.prompt([
+                        {
+                            type: 'confirm',
+                            name: 'isPurgeDeployFolder',
+                            message:
+                                'Do you want to empty the deploy folder (ensures no files from previous deployments remain)?',
+                            default: true,
+                        },
+                    ]);
+                }
+                if (skipInteraction || responses.isPurgeDeployFolders) {
+                    // Clear output folder structure for selected sub-type
+                    File.removeSync(File.normalizePath([properties.directories.deploy]));
+                }
+            }
+            const bdPromises = [];
+            for (const type in typeDelta) {
+                Util.logger.info(
+                    `⚡ mcdev bdb ${targetMlName} ${type} "${typeDelta[type].join(',')}"`
+                );
+                // omitting "await" to speed up creation
+                bdPromises.push(
+                    Builder.buildDefinitionBulk(targetMlName, type, typeDelta[type].join(','))
+                );
+            }
+            await Promise.all(bdPromises);
+            Util.logger.info(`- ✔️  Deploy defintions created`);
+            if (
+                properties.directories.templateBuilds == properties.directories.deploy ||
+                (Array.isArray(properties.directories.templateBuilds) &&
+                    properties.directories.templateBuilds.includes(properties.directories.deploy))
+            ) {
+                Util.logger.info(`You can now run deploy on the prepared BUs`);
+            } else {
+                Util.logger.info(
+                    `Your templated defintions are now ready to be copied into the deploy folder. Hint: You can have this auto-copied if you adjust directories.templateBuilds in your config.`
+                );
+            }
+        }
+    },
+
     /**
      * create markdown file for deployment listing
      *
